+++ src/components/navbar/leftNavBar.jsx
@@ -4,6 +4,7 @@ import PropType from "prop-types";
 import { makeStyles } from "@material-ui/core/styles";
 import {
   BookOutlined,
+  Cached,
   DescriptionTwoTone,
   Edit,
   HomeTwoTone,
@@ -39,11 +40,16 @@ const icons = [
     icon: <VpnKey />,
     name: "sign-in",
   },
+  {
+    icon: <Cached />,
+    isDisabledable: true,
+    name: "load-all",
+  },
 ];
 
-const propTypes = { onClick: PropType.func };
+const propTypes = { isLoading: PropType.bool, onClick: PropType.func };
 const useStyles = makeStyles(() => styles);
-const LeftNavBar = ({ onClick }) => {
+const LeftNavBar = ({ isLoading = false, onClick = null }) => {
   const classes = useStyles();
   return (
     <Fragment>
@@ -55,6 +61,7 @@ const LeftNavBar = ({ onClick }) => {
                 <Fragment key={icon.name}>
                   <NavIcon
                     icon={icon.icon}
+                    isDisabled={icon.isDisabledable && isLoading}
                     name={icon.name}
                     onClick={(event) => handleFunction(onClick, event)}
                   />