+++ src/sections/books/bookSection.jsx
@@ -1,18 +1,12 @@
 import React, { Fragment, useEffect, useState } from "react";
 import { connect } from "react-redux";
 import { useHistory } from "react-router-dom";
-import { v4 as uuidv4 } from "uuid";
 import PropType from "prop-types";
 import { Button } from "@material-ui/core";
 import RouteTitle from "../../components/routeTitle";
 import { BookCard, BookTitleSection } from "../../components/displays/book";
 import StandardButton from "../../components/standardButton";
-import { generateDateString } from "../../libs/date";
-import { downloadJson } from "../../libs/download";
-import { handleFunction } from "../../libs/eventUtil";
 import { sortByCreatedDate as sortFunc } from "../../libs/sorts";
-import { upsertNote } from "../../state/editor/actions";
-import { fetchArticles } from "../../state/notes/actions";
 
 const styleButton = { margin: "12px" };
 const styleHeader = { display: "inline-block", marginRight: "12px" };
@@ -20,22 +14,17 @@ const styleInline = { display: "inline-block" };
 const textHideEntires = "Hide Entries";
 const textShowEntries = "Show Entries";
 
-let abortCtrlFetchAll = null;
 const propTypes = {
-  getAllNotes: PropType.func,
   isLoading: PropType.bool,
   isUserSecured: PropType.bool,
   notes: PropType.array,
   pageTitle: PropType.string,
-  requestNoteUpsert: PropType.func,
 };
 const BookSection = ({
-  getAllNotes = null,
   isLoading = false,
   isUserSecured = false,
   notes = [],
   pageTitle = "",
-  requestNoteUpsert = null,
 }) => {
   const history = useHistory();
   const [expandEntry, setExpandEntry] = useState({});
@@ -43,12 +32,6 @@ const BookSection = ({
   const [showEntry, setShowEntry] = useState(true);
   const [showResources, setShowResources] = useState(false);
   const [updatedNotesKeys, setUpdatedNotesKeys] = useState({});
-
-  useEffect(() => {
-    return () => {
-      abortCtrlFetchAll?.abort();
-    };
-  }, []);
 
   useEffect(() => {
     const clonedNotes = JSON.parse(JSON.stringify(notes || [])).sort(sortFunc);
@@ -90,37 +73,12 @@ const BookSection = ({
     setExpandEntry(open);
   }, [notes]);
 
-  const handleClickAlignBookId = (tempBookId, noteId) => {
-    const aligningNote = notes.find((note) => {
-      return note.id === noteId;
-    });
-    aligningNote.bookId = meta[tempBookId].bookId;
-    aligningNote.updatedDate = generateDateString();
-    downloadJson(aligningNote, aligningNote.id || "data");
-    requestNoteUpsert(aligningNote, false);
-    setUpdatedNotesKeys((prev) => {
-      return { ...prev, [noteId]: true };
-    });
-  };
-
   const handleClickCardBtn = (id, type = "") => {
     if (!!id && ["clone", "edit"].indexOf(type) > -1) {
       const search = `id=${id}`;
       const pathname = `/${type}`;
       history.push({ pathname, search });
     }
-  };
-
-  const handleClickGenerateId = (id) => {
-    const clonedMeta = JSON.parse(JSON.stringify(meta));
-    clonedMeta[id].bookId = uuidv4();
-    setMeta(clonedMeta);
-  };
-
-  const handleClickLoadAll = () => {
-    abortCtrlFetchAll = new AbortController();
-    const config = { signal: abortCtrlFetchAll.signal };
-    handleFunction(getAllNotes, config);
   };
 
   const handleClickShowEntries = (id) => {
@@ -140,13 +98,6 @@ const BookSection = ({
       <section>
         <h4 style={styleHeader}>{`Total Books:`}</h4>
         <span style={styleInline}>{Object.keys(meta).length || `0`}</span>
-        <div style={{ ...styleInline, margin: 12 }}>
-          <StandardButton
-            disabled={isLoading}
-            label="Load All"
-            onClick={handleClickLoadAll}
-          />
-        </div>
         <div style={{ ...styleInline, margin: 12 }}>
           <StandardButton
             disabled={isLoading}
@@ -197,7 +148,6 @@ const BookSection = ({
                         isClonable={isUserSecured}
                         isEditable={isUserSecured}
                         noteData={data}
-                        onClickBookId={() => handleClickAlignBookId(bookId, id)}
                         onClickClone={() => handleClickCardBtn(id, "clone")}
                         onClickEdit={() => handleClickCardBtn(id, "edit")}
                       />
@@ -225,8 +175,5 @@ const mapStateToProps = (state) => ({
   isUserSecured: !!state.auth.token,
   notes: state.notes.notes,
 });
-const mapDispatchToProps = {
-  getAllNotes: fetchArticles,
-  requestNoteUpsert: upsertNote,
-};
+const mapDispatchToProps = {};
 export default connect(mapStateToProps, mapDispatchToProps)(BookSection);