+++ src/sections/notes/displayAllSection.jsx
@@ -10,12 +10,9 @@ import { FilterSortOrder, FilterTagSelector } from "../../components/filters";
 import { ArticleCard } from "../../components/displays/article";
 import { BookCard } from "../../components/displays/book";
 import RouteTitle from "../../components/routeTitle";
-import StandardButton from "../../components/standardButton";
-import { handleFunction } from "../../libs/eventUtil";
 import { generateTagsFromList } from "../../libs/generateTagList";
 import { journalTypes } from "../../libs/types";
 import * as _sorts from "../../libs/sorts";
-import { fetchArticles } from "../../state/notes/actions";
 import { notesGridStyles as styles } from "./styles";
 
 const orderTypes = {
@@ -24,9 +21,7 @@ const orderTypes = {
   title: "title",
 };
 
-let abortCtrlFetchAll = null;
 const propTypes = {
-  getAllNotes: PropType.func,
   isLoadingIndex: PropType.bool,
   isLoadingNotes: PropType.bool,
   isUserSecured: PropType.bool,
@@ -36,7 +31,6 @@ const propTypes = {
 };
 const useStyles = makeStyles(() => styles);
 const DisplayAllSection = ({
-  getAllNotes = null,
   isLoadingIndex = false,
   isLoadingNotes = false,
   isUserSecured = false,
@@ -157,13 +151,6 @@ const DisplayAllSection = ({
     }
   };
 
-  const handleClickLoadAll = () => {
-    abortCtrlFetchAll?.abort();
-    abortCtrlFetchAll = new AbortController();
-    const config = { signal: abortCtrlFetchAll.signal };
-    handleFunction(getAllNotes, config);
-  };
-
   const handleClickRemoveCurrentSelectedFilter = () => {
     if (!!filterTagSelected && !!filterTagSelected[0]) {
       setTagsFilter([
@@ -230,15 +217,6 @@ const DisplayAllSection = ({
             </Grid>
           </Grid>
         )}
-        {pageName === "view" && (
-          <div style={{ margin: 12 }}>
-            <StandardButton
-              disabled={isLoading}
-              label="Load All"
-              onClick={handleClickLoadAll}
-            />
-          </div>
-        )}
         {!isLoading &&
           Array.isArray(notes) &&
           notes
@@ -286,5 +264,5 @@ const mapStateToProps = (state) => ({
   isLoadingNotes: state.notes.isLoadingNotes,
   isUserSecured: !!state.auth.token,
 });
-const mapDispatchToProps = { getAllNotes: fetchArticles };
+const mapDispatchToProps = {};
 export default connect(mapStateToProps, mapDispatchToProps)(DisplayAllSection);