+++ src/components/navbar/index.js
@@ -1,3 +1,4 @@
 import LeftNavBar from "./leftNavBar";
+import NavIcon from "./navIcon";
 
-export { LeftNavBar };
+export { LeftNavBar, NavIcon };