+++ src/sections/books/bookSection.jsx
@@ -2,12 +2,15 @@ import React, { Fragment, useEffect, useState } from "react";
 import { connect } from "react-redux";
 import { useHistory } from "react-router-dom";
 import PropType from "prop-types";
-import { Button } from "@material-ui/core";
+import { Button, FormControlLabel, Switch } from "@material-ui/core";
 import RouteTitle from "../../components/routeTitle";
 import { BookCard, BookTitleSection } from "../../components/displays/book";
-import StandardButton from "../../components/standardButton";
+import { NavIcon } from "../../components/navbar";
 import { sortByCreatedDate as sortFunc } from "../../libs/sorts";
-
+import {
+  ArrowDropDown as DownArrow,
+  ArrowDropUp as UpArrow,
+} from "@material-ui/icons";
 const styleButton = { margin: "12px" };
 const styleHeader = { display: "inline-block", marginRight: "12px" };
 const styleInline = { display: "inline-block" };
@@ -99,10 +102,17 @@ const BookSection = ({
         <h4 style={styleHeader}>{`Total Books:`}</h4>
         <span style={styleInline}>{Object.keys(meta).length || `0`}</span>
         <div style={{ ...styleInline, margin: 12 }}>
-          <StandardButton
+          <FormControlLabel
+            control={
+              <Switch
+                checked={showEntry}
+                color={"primary"}
+                onChange={handleClickToggleDisplay}
+              />
+            }
             disabled={isLoading}
-            label="Toggle Display"
-            onClick={handleClickToggleDisplay}
+            label={`Showing ${showEntry ? "entries" : "resources"}`}
+            style={{ padding: "2px" }}
           />
         </div>
       </section>
@@ -110,55 +120,66 @@ const BookSection = ({
         const entries = meta[bookId].entryIds.length || 0;
         const isChild = entries > 1;
         return (
-          <section key={bookId}>
-            <BookTitleSection
-              author={meta[bookId].author}
-              bookDescription={meta[bookId].bookDescription}
-              bookSource={meta[bookId].bookSource}
-              pageCount={meta[bookId].pageCount}
-              publishDate={meta[bookId].publishDate}
-              publisher={meta[bookId].publisher}
-              title={meta[bookId].title}
-            />
-            <Button
-              onClick={() => handleClickShowEntries(bookId)}
-              size="small"
-              style={styleButton}
-              variant="outlined"
-            >
-              {expandEntry[bookId] ? textHideEntires : textShowEntries}
-              {`: ${entries}`}
-            </Button>
+          <section key={bookId} style={{ paddingBottom: "6px" }}>
+            <div style={{ display: "flex" }}>
+              <div style={{ display: "inline-block", marginRight: "12px" }}>
+                <div
+                  style={{ display: "block", width: "72px" }}
+                >{`Notes: ${entries}`}</div>
+                <div style={{ display: "block" }}>
+                  <NavIcon
+                    icon={expandEntry[bookId] ? <UpArrow /> : <DownArrow />}
+                    name={expandEntry[bookId] ? "expand" : "collapse"}
+                    onClick={() => handleClickShowEntries(bookId)}
+                  />
+                </div>
+              </div>
+              <div style={{ display: "inline-block", overflowX: "hidden" }}>
+                <BookTitleSection
+                  author={meta[bookId].author}
+                  bookDescription={meta[bookId].bookDescription}
+                  bookSource={meta[bookId].bookSource}
+                  pageCount={meta[bookId].pageCount}
+                  publishDate={meta[bookId].publishDate}
+                  publisher={meta[bookId].publisher}
+                  title={meta[bookId].title}
+                />
+              </div>
+            </div>
+            {expandEntry[bookId] && showResources && (
+              <ul>
+                {meta[bookId].entryIds.map((id) => {
+                  const data = notes.find((note) => {
+                    return note.id === id;
+                  });
+                  return data.resources?.map((res, idx) => {
+                    return <li key={res.id || idx}>{res.resource}</li>;
+                  });
+                })}
+              </ul>
+            )}
             {expandEntry[bookId] &&
+              showEntry &&
               meta[bookId].entryIds.map((id) => {
                 const data = notes.find((note) => {
                   return note.id === id;
                 });
                 return (
                   <Fragment key={id}>
-                    {showEntry && (
-                      <BookCard
-                        hasBorder={true}
-                        isButtonMisc={
-                          !updatedNotesKeys[id] &&
-                          isUserSecured &&
-                          data.bookId !== meta[bookId].bookId
-                        }
-                        isChild={isChild}
-                        isClonable={isUserSecured}
-                        isEditable={isUserSecured}
-                        noteData={data}
-                        onClickClone={() => handleClickCardBtn(id, "clone")}
-                        onClickEdit={() => handleClickCardBtn(id, "edit")}
-                      />
-                    )}
-                    {showResources && (
-                      <ul>
-                        {data.resources?.map((res, idx) => {
-                          return <li key={res.id || idx}>{res.resource}</li>;
-                        })}
-                      </ul>
-                    )}
+                    <BookCard
+                      hasBorder={true}
+                      isButtonMisc={
+                        !updatedNotesKeys[id] &&
+                        isUserSecured &&
+                        data.bookId !== meta[bookId].bookId
+                      }
+                      isChild={isChild}
+                      isClonable={isUserSecured}
+                      isEditable={isUserSecured}
+                      noteData={data}
+                      onClickClone={() => handleClickCardBtn(id, "clone")}
+                      onClickEdit={() => handleClickCardBtn(id, "edit")}
+                    />
                   </Fragment>
                 );
               })}